requirements.txt
discord.py>=2.3.0
python-dotenv==1.0.0
aiohttp>=3.9.0
# audioop
psutil>=5.9.0
pyotp>=2.9.0
qrcode[pil]>=7.4.0
# asyncpg==0.29.0